+++ lib/table.ts
@@ -1,68 +1,112 @@
 import {generateSelect} from './select'
 import {generateFrom} from './from'
-import {joinNonNullWithNewLine} from './parsing'
+import {joinNonNullWithNewLine, joinWithNewLine} from './parsing'
 import {generateMap} from './map'
 import {generateFilter} from './filter'
+import {generateSortBy} from './sort_by'
 
 export interface Constructor<T> {
     new (...args: any[]): T
 }
 
 class SelectTable<T> {
-    private readonly selectSql: string
-
-    constructor(private table: Table<T>) {
-        this.selectSql = generateSelect(this.table.constructor)
-    }
+    constructor(
+        private readonly generateTableSql: () => string,
+        private readonly ctor: Constructor<T>) {}
 
     toString(): string {
-        return joinNonNullWithNewLine([
-            this.selectSql,
-            this.table.fromSql,
-            this.table.filterSql
+        return joinWithNewLine([
+            generateSelect(this.ctor),
+            this.generateTableSql()
         ])
     }
 }
 
 class MapTable<T, U> {
-    private readonly mapSql: string
-
-    constructor(private table: Table<T>, private map: (x: T) => U) {
-        this.mapSql = generateMap(this.map)
-    }
+    constructor(
+        private readonly generateTableSql: () => string,
+        private readonly map: (x: T) => U) {}
 
     toString(): string {
-        return joinNonNullWithNewLine([
-            this.mapSql,
-            this.table.fromSql,
-            this.table.filterSql
+        return joinWithNewLine([
+            generateMap(this.map),
+            this.generateTableSql()
         ])
     }
 }
 
+export type By<T> = (x: T) => number|string
+
+export type Order<T> = {
+    by: By<T>
+    direction: 'asc'|'desc'
+}
+
 export class Table<T> {
-    public readonly fromSql: string
-    public readonly filterSql: string
-
     constructor(
-        public readonly constructor: Constructor<T>,
-        public readonly name: string,
-        public readonly predicates: Array<(x: T) => boolean> = []) {
-
-        this.fromSql = generateFrom(this.name)
-        this.filterSql = this.predicates.length > 0 ? generateFilter(this.predicates) : null
+        private readonly constructor: Constructor<T>,
+        private readonly name: string,
+        private readonly predicates: Array<(x: T) => boolean> = []) {
     }
 
     filter(predicate: (x: T) => boolean): Table<T> {
         return new Table(this.constructor, this.name, this.predicates.concat(predicate))
     }
 
+    sortBy(by: By<T>): SortTable<T> {
+        return new SortTable(() => this.toString(), this.constructor, [{ by, direction: 'asc'}])
+    }
+
+    sortDescendinglyBy(by: By<T>): SortTable<T> {
+        return new SortTable(() => this.toString(), this.constructor, [{ by, direction: 'desc'}])
+    }
+
     select(): SelectTable<T> {
-        return new SelectTable(this)
+        return new SelectTable(() => this.toString(), this.constructor)
     }
 
     map<U>(f: (x: T) => U): MapTable<T, U> {
-        return new MapTable(this, f)
+        return new MapTable(() => this.toString(), f)
+    }
+
+    toString(): string {
+        const fromSql = generateFrom(this.name)
+        const filterSql = this.predicates.length > 0 ? generateFilter(this.predicates) : null
+
+        return joinNonNullWithNewLine([
+            fromSql,
+            filterSql])
+    }
+}
+
+export class SortTable<T> {
+    constructor(
+        private readonly generateTableSql: () => string,
+        private readonly ctor: Constructor<T>,
+        private readonly orders: Array<Order<T>>) {
+    }
+
+    thenBy(by: By<T>): SortTable<T> {
+        return new SortTable(this.generateTableSql, this.ctor, this.orders.concat({by, direction: 'asc'}))
+    }
+
+    thenDescendinglyBy(by: By<T>): SortTable<T> {
+        return new SortTable(this.generateTableSql, this.ctor, this.orders.concat({by, direction: 'desc'}))
+    }
+
+    select(): SelectTable<T> {
+        return new SelectTable(() => this.toString(), this.ctor)
+    }
+
+    map<U>(f: (x: T) => U): MapTable<T, U> {
+        return new MapTable(() => this.toString(), f)
+    }
+
+    toString(): string {
+        return joinWithNewLine([
+            this.generateTableSql(),
+            generateSortBy(this.orders)
+        ])
     }
 }