+++ test/select.test.ts
@@ -4,6 +4,6 @@ import * as assert from 'assert'
 
 describe('parseSelect', () => {
     it('returns a comma/space-separated list of aliased field names', () => {
-        assert.equal(computeFieldList(Employee), 't1.id, t1.first_name, t1.last_name, t1.title, t1.department_id')
+        assert.equal(computeFieldList(Employee), 't1.id, t1.first_name, t1.last_name, t1.title, t1.salary, t1.department_id')
     })
 })