+++ test/test_tables.ts
@@ -6,6 +6,7 @@ export class Employee {
         public firstName: string,
         public lastName: string,
         public title: string,
+        public salary: number,
         public departmentId: string) {
     }
 }