+++ Changelog.md
@@ -10,3 +10,8 @@
 - Introduced changelog
 - Described query support in the readme
 - Started to use SQL syntax in the readme 
+
+## v0.2
+
+- Implemented basic support for sorting by multiple orders in ascending and descending direction
+- Added sorting examples to the readme